+++ front-end-sistema-departamento/src/js/role/roleRemove.js
@@ -1,4 +1,11 @@
 function roleRemove(id) {
+
+    if (!loginVerification()) {
+        alert("Você não está logado")
+        window.location.href = "../../index.html"
+        return false
+    }
+
     if (confirm("Deseja realmente deletar esse cargo?")) {
         axios.delete(`${domain}/roles/${id}`, header)
             .then(function(response) {
@@ -8,7 +15,7 @@ function roleRemove(id) {
             })
             .catch(function(error) {
                 // handle error
-                alert(error)
+                console.log(error)
             })
     }
 }
